FEAT: Add router to manage page transition
