+++ src/components/modules/Header.tsx
@@ -9,13 +9,13 @@ import Toolbar from '@mui/material/Toolbar'
 import Typography from '@mui/material/Typography'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faUser } from '@fortawesome/free-solid-svg-icons'
+import { useRouter } from 'next/router'
 
 import LoginForm from './LoginForm'
 import { useAuthentication } from 'hooks/firebase/useAuthentication'
-import { StepperHandlerContext } from 'components/pages/PlaningMain'
 
 const Header = () => {
-  const setStep = React.useContext(StepperHandlerContext)
+  const router = useRouter()
 
   const [user, auth] = useAuthentication()
   const [open, setOpen] = React.useState<'signIn' | 'signUp' | null>(null)
@@ -40,7 +40,7 @@ const Header = () => {
         <Toolbar variant="dense">
           <Button
             color="inherit"
-            onClick={() => setStep('Home')}
+            onClick={() => router.push('/')}
             sx={{ mr: 1 }}>
             Home
           </Button>
